requirements.txt
# Web framework
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
jinja2>=3.1.0
python-multipart>=0.0.9
itsdangerous>=2.2.0

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0

# Auth
authlib>=1.3.0
httpx>=0.27.0

# Environment
python-dotenv>=1.0.0

# Import
pandas>=2.2.0
openpyxl>=3.1.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Code quality
black>=24.0.0
ruff>=0.4.0
